added desc. to .py files + separated slurm classes/fncs in workload manager + created validation service for all arg validations
