+++ ga_dashboard/backend/services/sacct_service.py
@@ -1,7 +1,15 @@
+# ------------------------------------------------------------------
+# Service to interact with the SLURM workload manager using the 'sacct' command.
+# ------------------------------------------------------------------
+
 import subprocess
 import pandas as pd
 
 class SacctService:
+    """
+    Service to interact with the SLURM workload manager using the 'sacct' command.
+    Contains separate methods to pull logs in different contexts (by time, by JobID, etc.) and can be extended with more methods as needed.
+    """
 
     bash_com = [
                 "sacct",
